requirements.txt
# python-telegram-bot[job-queue,webhooks]==20.7
python-telegram-bot==13.15